Finish first version of optimized LLR generation subassembly
This creates filtered LLR terms using only four jobs. There's more work
to be done, specifically around optimizing the term collection process
initially, but good enough for now.

+++ src/main/java/com/scaleunlimited/cascading/ml/ITermsFilter.java
@@ -2,7 +2,7 @@ package com.scaleunlimited.cascading.ml;
 
 public interface ITermsFilter {
 
-    public boolean filter(float llrScore, String term, ITermsParser parser);
+    public boolean filter(double llrScore, String term, ITermsParser parser);
     
     public int getMaxResults();
 }
